Remove Number module and move Number things to Algebra/Adapters
